blog/requirements.txt
certifi==2021.5.30
chardet==4.0.0
Django==2.2.4
django-bootstrap-pagination==1.7.1
django-ckeditor==6.1.0
django-embed-video==1.4.0
django-js-asset==1.2.2
django-social-share==2.2.1
idna==2.10
Pillow==8.3.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
